add optional jwt

+++ lesson_11/resources/item.py
@@ -1,4 +1,4 @@
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt, get_jwt_identity
 from flask_restful import Resource, reqparse
 
 from models.item import ItemModel
@@ -42,6 +42,10 @@ class Item(Resource):
 
     @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message': 'You don\'t have enough power to do this'}
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -49,6 +53,11 @@ class Item(Resource):
 
 
 class ItemList(Resource):
-    @jwt_required()
+    @jwt_required(optional=True)
     def get(self):
-        return {'items': [item.json() for item in ItemModel.get_all_items()]}, 200
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.get_all_items()]
+        if user_id:
+            return {'items': items}, 200
+        else:
+            return {'items': [item['name'] for item in items]}, 200
